Split the tests to mirror the modules
One file per module, so a change to loading.py has one obvious test file and
two people editing different modules no longer edit the same test file:

  helpers.py          _comp, _source_for and the CSV fixtures
  test_model.py       Component and index_by_id
  test_colors.py      the palette and owner-colour loading
  test_loading.py     cell parsing and everything load_components does
  test_naming.py      SVG ids, clone ids, layer filename stems
  test_validation.py  reference and duplicate checks
  test_render.py      node attributes, externals, edge styles
  test_export.py      components.json
  test_cli.py         the .env precedence for --google-sheet

The id-collision tests stay together in test_naming.py even though they call
validate(): the namespace they protect is naming.py's contract, and splitting
them across two files would undo the grouping they were given a commit ago.
The file says so at the top.

Same 108 tests, same names. Legend has no tests today; when it gets some they
join test_render.py rather than starting a tenth file.
